自动提取关键词的工具等

+++ src/main/java/com/tktick/dao/repository/TkArticleRepository.java
@@ -3,12 +3,13 @@ package com.tktick.dao.repository;
 import org.springframework.stereotype.Repository;
 import com.tktick.bean.entity.TkArticle;
 import com.tktick.dao.BaseRepository;
+
 /**
  * JpaSpecificationExecutor 方便分页
  * @author user
  *
  */
 @Repository
-public interface TkArticleRepository extends BaseRepository<TkArticle, Long> {
+public interface TkArticleRepository extends BaseRepository<TkArticle, Integer> {
 	
 }
